Replace Paynote by PaynoteV2
