+++ components/Article/Page.js
@@ -11,7 +11,7 @@ import SeriesNavButton from './SeriesNavButton'
 import PdfOverlay, { getPdfUrl, countImages } from './PdfOverlay'
 import Extract from './Extract'
 import withT from '../../lib/withT'
-import { PayNote, MAX_PAYNOTE_SEED } from './PayNoteV2'
+import { PayNote, MAX_PAYNOTE_SEED } from './PayNote'
 import withInNativeApp, { postMessage } from '../../lib/withInNativeApp'
 import { cleanAsPath } from '../../lib/routes'
 import { createRequire } from '@project-r/styleguide/lib/components/DynamicComponent'
